% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliced_wasserstein.R
\name{compute_all_distances}
\alias{compute_all_distances}
\title{Pairwise Wasserstein Distances Along Projection Directions Between Several Empirical Distributions}
\usage{
compute_all_distances(
  distributions,
  thetas,
  A = NULL,
  verbose = TRUE,
  keep_projections = TRUE
)
}
\arguments{
\item{distributions}{A list of matrices representing empirical distributions.}

\item{thetas}{A matrix, each row of which represents a projection direction.}

\item{A}{Optionally, a matrix used to transform each projection direction.}

\item{verbose}{If \code{TRUE}, show progress.}

\item{keep_projections}{If \code{TRUE}, the distance matrix for each projection direction is output.
If \code{FALSE}, the distance matrices for the different projection directions are averaged.}
}
\value{
If \code{keep_projections = TRUE}, a list of squared-distance matrices, one for each projection direction;
otherwise, a matrix with the averaged squared distances.
}
\description{
This function computes the squared 2-Wasserstein distances between the projections
of several empirical directions along specified projection directions.
The projection directions can optionally be transformed by means of a linear map.
The functions can output the squared distances along each projection direction,
or the squared distances averaged over projection directions.
}
\examples{
M1 <- matrix(rnorm(50), ncol = 5)
M2 <- matrix(rnorm(150), ncol = 5)
M3 <- matrix(rnorm(250), ncol = 5)
# Sliced Wasserstein:
my_directions <- generate_directions(20, 5)
compute_all_distances(list(M1, M2, M3), my_directions, keep_projections = FALSE)
# Marginal Wasserstein distances:
marginal_wass <- compute_all_distances(list(M1, M2, M3), diag(1, 5), keep_projections = TRUE)
marginal_wass[[3]] # along third dimension
}
