% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLICER.R
\name{detect_cell_types}
\alias{detect_cell_types}
\title{Identify clusters corresponding to putative cell types}
\usage{
detect_cell_types(embedding, k)
}
\arguments{
\item{embedding}{Low-dimensional LLE embedding of cells}

\item{k}{Number of nearest neighbors to use when detecting clusters}
}
\value{
Vector containing a numerical cluster assignment for each cell
}
\description{
\code{detect_cell_types} divides the k-nearest neighbor
graph (built from the LLE embedding) into connected
components. These connected components represent clusters of cells corresponding to putative cell types.
}
