% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dot}
\alias{dot}
\title{Compute dot products}
\usage{
dot(a, b)
}
\arguments{
\item{a}{A numeric vector.}

\item{b}{A numeric vector or matrix.}
}
\value{
If \code{b} is a vector, the dot product of \code{a} and \code{b}.
If \code{b} is a matrix, a vector containing the dot products of \code{a} and each row of \code{b}.
}
\description{
Computes the dot product between a vector and one or several other vectors.
}
\examples{
x <- c(1, 2, 3)
y <- c(2, 4, 2)
dot(x, y)
m <- rbind(x, y)
dot(x, m)
}
