% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_processes.R
\name{fit_gpr}
\alias{fit_gpr}
\title{Gaussian Process Fitting to Test Data with Tuned Hyperparameters}
\usage{
fit_gpr(
  D2,
  training_idx,
  test_idx,
  y_train,
  y_test = NULL,
  centre = TRUE,
  use_gradient = TRUE,
  runs = 10,
  verbose = TRUE
)
}
\arguments{
\item{D2}{A matrix containing squared pairwise distances between all objects.}

\item{training_idx}{A vector with the row (and column) numbers corresponding to the training entries in \code{D2}.}

\item{test_idx}{A vector with the row (and column) numbers corresponding to the test entries in \code{D2}.}

\item{y_train}{A vector with the training outcomes.}

\item{y_test}{Optionally, a vector with the test outcomes.}

\item{centre}{If \code{TRUE}, the training outcomes are centred around their mean before fitting the model. The mean is then added back to the predictions at the end.}

\item{use_gradient}{If \code{TRUE}, closed-form expressions for the RBF's gradient are used.
Else, the optimiser uses the finite-differences method.}

\item{runs}{Number of independent attempts to find a minimum when optimising the hyperparameters.}

\item{verbose}{If \code{TRUE}, progress is shown on the console.}
}
\value{
A list containing the predictions for the test objects,
the root mean squared error (if the true test outcomes are provided),
and the three tuned hyperparameter values.
}
\description{
This function generates predictions using a Gaussian process model
with a Gaussian RBF kernel. The hyperparameters are tuned using the
training data by minimising the model's negative marginal log-likelihood.
}
\examples{
N1 <- 25
N2 <- 10
x_train <- seq(-pi, pi, length.out = N1)
x_test  <- runif(N2, -pi, pi)
y_train <- x_train * plogis(x_train) * cos(x_train)
y_test <- x_test * plogis(x_test) * cos(x_test)
D2 <- outer(c(x_train, x_test), c(x_train, x_test), "-")^2
fit <- fit_gpr(D2, seq_len(N1), N1 + seq_len(N2), y_train, y_test, runs = 50)
curve(x * plogis(x) * cos(x), -pi, pi)
points(x_train, y_train, pch = 1)
points(x_test, fit$test_predictions, pch = 16)
fit$RMSE
}
