% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_processes.R
\name{fit_gpr_multiple}
\alias{fit_gpr_multiple}
\title{Gaussian Process Fitting to Test Data Using Multiple Kernels with Tuned Hyperparameters}
\usage{
fit_gpr_multiple(
  D2_list,
  training_idx,
  test_idx,
  y_train,
  y_test = NULL,
  centre = TRUE,
  use_gradient = TRUE,
  runs = 10,
  verbose = TRUE
)
}
\arguments{
\item{D2_list}{A list of matrices containing squared pairwise distances between all objects.}

\item{training_idx}{A vector with the row (and column) numbers corresponding to the training entries in \code{D2}.}

\item{test_idx}{A vector with the row (and column) numbers corresponding to the test entries in \code{D2}.}

\item{y_train}{A vector with the training outcomes.}

\item{y_test}{Optionally, a vector with the test outcomes.}

\item{centre}{If \code{TRUE}, the training outcomes are centred around their mean before fitting the model. The mean is then added back to the predictions at the end.}

\item{use_gradient}{If \code{TRUE}, closed-form expressions for the RBF's gradient are used.
Else, the optimiser uses the finite-differences method.}

\item{runs}{Number of independent attempts to find a minimum when optimising the hyperparameters.}

\item{verbose}{If \code{TRUE}, progress is shown on the console.}
}
\value{
A list containing the predictions for the test objects,
the root mean squared error (if the true test outcomes are provided),
and the tuned hyperparameter values.
}
\description{
This is the counterpart to \link{fit_gpr} when multiple squared-distance
matrices are available. A kernel is then built as a linear combination
of base Gaussian RBF kernels. The hyperparameters are tuned by minimising
the model's negative marginal log-likelihood on the training data.
}
\examples{
N1 <- 25
N2 <- 10
x_train1 <- runif(N1, -pi, pi)
x_train2 <- runif(N1, 0, 1)
x_test1  <- runif(N2, -pi, pi)
x_test2  <- runif(N2, 0, 1)
y_train <- x_train2 * plogis(x_train1) * cos(x_train1)
y_test <- x_test2 * plogis(x_test1) * cos(x_test1)
D2_1 <- outer(c(x_train1, x_test1), c(x_train1, x_test1), "-")^2
D2_2 <- outer(c(x_train2, x_test2), c(x_train2, x_test2), "-")^2
fit <- fit_gpr_multiple(list(D2_1, D2_2), seq_len(N1), N1 + seq_len(N2), y_train, y_test, runs = 50)
plot(fit$test_predictions, y_test)
fit$RMSE
}
