% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_directions.R
\name{generate_directions}
\alias{generate_directions}
\title{Generate random projection directions}
\usage{
generate_directions(L, d)
}
\arguments{
\item{L}{Number of projection directions.}

\item{d}{Dimension.}
}
\value{
An \code{L} by \code{d} matrix, each row of which contains a unit vector.
}
\description{
Obtain \code{L} random vectors on the unit sphere in \code{d}-dimensional Euclidean space.
}
\examples{
generate_directions(10, 4)
plot(generate_directions(25, 2))
}
