% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLICER.R
\name{graph_process_distance}
\alias{graph_process_distance}
\title{Plot trajectory colored by process distance}
\usage{
graph_process_distance(traj_graph, embedding, start, cell_symbols = 16)
}
\arguments{
\item{traj_graph}{Nearest neighbor graph built from LLE embedding}

\item{embedding}{Low-dimensional LLE embedding of cells}

\item{start}{Index of start cell}

\item{cell_symbols}{Symbols to use for plotting each cell}
}
\value{
None
}
\description{
This function plots the embedding produced by LLE, coloring
cells by their progress through a process.
}
\examples{
\dontrun{
traj_lle = lle(traj[,genes],m=2,k)
traj_graph = conn_knn_graph(traj_lle,5)
find_extreme_cells(traj_graph)
graph_process_distance(traj_graph,start)
}
}

