% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_sum}
\alias{kernel_sum}
\title{Sum of Kernels}
\usage{
kernel_sum(D2_list, length_scales, variances, kernels)
}
\arguments{
\item{D2_list}{A list of matrices with squared pairwise distances.}

\item{length_scales}{A vector of length-scale parameters.}

\item{variances}{A vector of kernel variance parameters.}

\item{kernels}{A vector of kernels (\code{rbf}, \code{matern05}, \code{matern15}, \code{matern25}).}
}
\value{
A kernel matrix with the same dimensions as the matrices in \code{D_list}.
}
\description{
Builds kernel matrices from a list of distances matrices and computes
their sum.
}
\examples{
M1 <- rbind(c(0, 2, 1),
            c(2, 0, 2.5),
            c(1, 2.5, 0))
M2 <- rbind(c(0, 1, 4),
            c(1, 0, 3),
            c(4, 3, 0))
M_list <- list(M1^2, M2^2)
kernel_sum(M_list, length_scales = c(1, 2), variances = c(2, 1/2),
  kernels = c("rbf", "matern15"))

}
