% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{matern}
\alias{matern}
\title{Matérn Kernels}
\usage{
matern(D, length_scale = 1, variance = 1, tau = 1.5, squared = TRUE)
}
\arguments{
\item{D}{Matrix with pairwise (possibly squared) distances.}

\item{length_scale}{Length-scale parameter.}

\item{variance}{Kernel variance parameter.}

\item{tau}{Smoothness parameter. needs to be 0.5, 1.5 (default) or 2.5.}

\item{squared}{If \code{TRUE} (default), \code{D} is assumed to contain squared
distances. Set to \code{FALSE} if this is not the case.}
}
\value{
A kernel matrix with the same dimensions as \code{D}.
}
\description{
Builds a Matérn kernel (with smoothness parameter 0.5, 1.5 or 2.5) from a
distance matrix.
}
\examples{
M <- rbind(c(0, 2, 1),
           c(2, 0, 2.5),
           c(1, 2.5, 0))
matern(M, tau = 0.5, squared = FALSE)
matern(M, tau = 1.5, squared = FALSE)
matern(M, tau = 2.5, squared = FALSE)
}
