% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{matern_multiple}
\alias{matern_multiple}
\title{Sum of Matérn Kernels}
\usage{
matern_multiple(D_list, length_scales, variances, taus, squared = TRUE)
}
\arguments{
\item{D_list}{A list of matrices with pairwise (possibly squared) distances.}

\item{length_scales}{A vector of length-scale parameters.}

\item{variances}{A vector of kernel variance parameters.}

\item{taus}{A vector of smoothness parameters.
Values need to be in 0.5, 1.5 and 2.5.
If only a single value is provided, it will be used for all components.}

\item{squared}{If \code{TRUE} (default), each distance matrix is assumed to
contain squared distances. If \code{FALSE}, each distance matrix is assumed to
contain raw distances.}
}
\value{
A kernel matrix with the same dimensions as the matrices in \code{D_list}.
}
\description{
Builds Matérn kernel matrices from a list of distances matrices and computes
their sum.
}
\examples{
M1 <- rbind(c(0, 2, 1),
            c(2, 0, 2.5),
            c(1, 2.5, 0))
M2 <- rbind(c(0, 1, 4),
            c(1, 0, 3),
            c(4, 3, 0))
M_list <- list(M1, M2)
matern_multiple(M_list, length_scales = c(1, 2), variances = c(2, 1/2),
  taus = c(0.5, 2.5), squared = FALSE)
}
