% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d_optimal_transport.R
\name{nw_corner_distance}
\alias{nw_corner_distance}
\title{One-Dimensional Wasserstein Distance Using North-West Corner Algorithm}
\usage{
nw_corner_distance(
  x,
  y,
  presorted = FALSE,
  p = 2,
  eps = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x, y}{Vectors representing one-dimensional empirical distributions.}

\item{presorted}{Set to \code{TRUE} if both \code{x} and \code{y} are sorted to obtain a speed-up.}

\item{p}{Order of the Wasserstein distance.}

\item{eps}{Numerical precision for floating point comparisons.}
}
\value{
The p-Wasserstein distance between \code{x} and \code{y}.
}
\description{
Applies the north-west corner algorithm to obtain the p-Wasserstein distance
between two one-dimensional empirical distributions.
}
\examples{
x <- rnorm(10)
y <- rnorm(40)
nw_corner_distance(x, y)
nw_corner_distance(sort(x), sort(y), presorted = TRUE)
}
