% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{rbf}
\alias{rbf}
\title{Gaussian radial basis function kernel}
\usage{
rbf(D, length_scale = 1, variance = 1, squared = TRUE)
}
\arguments{
\item{D}{A matrix with pairwise (possibly squared) distances.}

\item{length_scale}{The length-scale parameter.}

\item{variance}{The kernel variance parameter.}

\item{squared}{If \code{TRUE} (default), \code{D} is assumed to contain squared distances.
Set to \code{FALSE} if this is not the case.}
}
\value{
A kernel matrix with the same dimensions as \code{D}.
}
\description{
Builds a Gaussian radial basis function kernel matrix from
a distance matrix.
}
\examples{
M <- rbind(c(0, 2, 1),
           c(2, 0, 2.5),
           c(1, 2.5, 0))
rbf(M, squared = FALSE)
}
