% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{rbf_multiple}
\alias{rbf_multiple}
\title{Sum of Gaussian radial basis function kernels}
\usage{
rbf_multiple(D_list, length_scales, variances, squared = TRUE)
}
\arguments{
\item{D_list}{A list of matrices with pairwise (possibly squared) distances.}

\item{length_scales}{A vector of length-scale parameters.}

\item{variances}{A vector of kernel variance parameters.}

\item{squared}{If \code{TRUE} (default), each distance matrix is assumed to contain squared distances.
If \code{FALSE}, each distance matrix is assumed to contain raw distances.}
}
\value{
A kernel matrix with the same dimensions as the matrices in \code{D_list}.
}
\description{
Builds Gaussian radial basis function kernel matrices from a list of
distances matrices and computes their sum.
}
\examples{
M1 <- rbind(c(0, 2, 1),
            c(2, 0, 2.5),
            c(1, 2.5, 0))
M2 <- rbind(c(0, 1, 4),
            c(1, 0, 3),
            c(4, 3, 0))
M_list <- list(M1, M2)
rbf_multiple(M_list, length_scales = c(1, 2), variances = c(2, 1/2), squared = FALSE)
}
