% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLICER.R
\name{select_genes}
\alias{select_genes}
\title{Select genes to use in building a cell trajectory}
\usage{
select_genes(embedding)
}
\arguments{
\item{embedding}{Low-dimensional LLE embedding of cells}
}
\value{
Vector containing indices of selected genes
}
\description{
This function uses "neighborhood variance" to identify genes
that vary smoothly, rather than fluctuating randomly, across
the set of cells. Genes selected in this way can then be used
to construct a trajectory.
}
\examples{
\dontrun{
genes = select_genes(traj)
}
}
