% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliced_wasserstein.R
\name{sliced_wasserstein}
\alias{sliced_wasserstein}
\title{Sliced Wasserstein Distance Between Two Empirical Distributions}
\usage{
sliced_wasserstein(x, y, p = 2, thetas = NULL, L = 50, seed = NULL)
}
\arguments{
\item{x, y}{Matrices representing empirical distributions.}

\item{p}{Order of sliced Wasserstein distance.}

\item{thetas}{Optionally, a matrix, each row of which represents a projection direction.}

\item{L}{If no \code{thetas} are provided, \code{L} random projection directions are generated.}

\item{seed}{Optional random seed.}
}
\value{
Estimated sliced Wasserstein distance of order \code{p} between \code{x} and \code{y}.
}
\description{
Estimate the sliced Wasserstein distance between two empirical distributions
represented as matrices using Monte Carlo simulation.

Optionally, a matrix \code{thetas} can be supplied. Each row of \code{thetas} will
be interpreted as a projection direction. These rows do not have to be
unit vectors, allowing users to transform the projection directions
rather than the matrices.
}
\examples{
M1 <- matrix(rnorm(50), ncol = 5)
M2 <- matrix(rnorm(250), ncol = 5)
sliced_wasserstein(M1, M2) # random projection directions
cardinal_axes <- diag(1, 5)
sliced_wasserstein(M1, M2, thetas = cardinal_axes)
first_two_axes <- cardinal_axes[1:2, ]
sliced_wasserstein(M1, M2, thetas = first_two_axes)
}
