% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{squared_norm}
\alias{squared_norm}
\title{Compute squared norms}
\usage{
squared_norm(a)
}
\arguments{
\item{a}{A numeric vector or matrix.}
}
\value{
The squared Euclidean norm of \code{a} (if \code{a} is a vector)
or the squared Euclidean norms of each row of \code{a} (if \code{a} is a matrix).
}
\description{
Compute squared norms
}
\examples{
x <- c(1, 2, 3)
squared_norm(x)
m <- matrix(rnorm(40), nrow = 10)
squared_norm(m)
}
