% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj.R
\docType{data}
\name{traj}
\alias{traj}
\title{This is a dataset containing a synthetic branching trajectory.}
\format{A matrix with samples down rows and genes across columns}
\usage{
data(traj)
}
\description{
This is a dataset containing a synthetic branching trajectory.
}
\keyword{datasets}
