% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLICER.R
\name{width_k}
\alias{width_k}
\title{Helper function for k selection}
\usage{
width_k(k, traj_exp)
}
\arguments{
\item{k}{Nearest neighbors}

\item{traj_exp}{Cell expression matrix}
}
\value{
Width of LLE embedding
}
\description{
Helper function for k selection
}
